+++ src/cipher/encryptingstorage/__init__.py
@@ -146,10 +146,10 @@ class EncryptingStorage(object):
         """
         return self.db.invalidateCache()
 
-    def invalidate(self, transaction_id, oids, version=''):  # the version arg is still present in ZODB 4.3.1, but is removed in ZODB 5.2.3
+    def invalidate(self, transaction_id, oids, version=''):
         """ For IStorageWrapper
         """
-        return self.db.invalidate(transaction_id, oids, version)
+        return self.db.invalidate(transaction_id, oids)
 
     def references(self, record, oids=None):
         """ For IStorageWrapper