+++ CHANGES.rst
@@ -5,6 +5,8 @@ CHANGES
 1.1.1 (unreleased)
 ------------------
 
+- Add support for ZODB/ZEO 5.
+
 - Close the underlying iterator used by the iterator wrapper when it is closed.
   (https://github.com/zopefoundation/zc.zlibstorage/issues/4)