+++ CHANGES.rst
@@ -10,7 +10,7 @@ CHANGES
 - Close the underlying iterator used by the iterator wrapper when it is closed.
   (https://github.com/zopefoundation/zc.zlibstorage/issues/4)
 
-- Add Python 3.4, 3.5, 3.6 compatibility.
+- Add Python 3.5, 3.6, 3.7, 3.8 and 3.9 compatibility.
 
 
 1.1 (2016-04-22)